added new files for camera, contour finding, spike position finding, and network tabling
