+++ SpikeFinding.py
@@ -1,0 +1,9 @@
+import cv2
+import numpy as np
+
+def findCenter(contours):
+	if len(contours) == 0:
+		return False
+	contour = np.concatenate(contours)
+	x,y,w,h = cv2.boundingRect(contour)
+	return (x + w/2, y + h/2)
